fix: generate auto-imports for more pattern completions
This tweaks the case-completion logic to always take auto-imports into
account. The auto-import logic was present for known direct subclasses,
but not for some other cases. Now, they all get handled uniformly.

The only hair in the soup is that extra work needs to go in to

   def myList: List[Int]
   myList match {
     case N@@   // want `case Nil =>`, not `case immutable.Nil =>`
   }

Reason here is that the `Nil` variant is completing with
`scala.collection.immutable.Nil`, while only `scala.Nil` is in scope.
The `scala.*` package re-exports are narrowly worked around since they
probably would occur frequently.