fix: colliding pattern and scope completions
This ensures that scope-based completions show up in a pattern context,
even if the symbol to complete is the same as the one underlying a
pattern completion (see the `underscore-and-object` test added)

This is achieved by putting case pattern completions in a new completion
member with its own dedup-ing `id`.

Although it turned out to not matter in this specific case, also fixed
`inverseSemanticdbSymbol` incorrectly returning an object symbol for
some class symbols. That tweak on its own didn't affect any test
output.