prefer `.bazelproject` as the Bazel BSP root
Noticed that in a monorepo Metals Bazel import was working for some
people and not others. It turns out that we have some secondary files
with the `.bazelproject` suffix and depending on the order of the
directory listing, you would get either the right `.bazelproject` file
or else some `.personal.bazelproject` file (incidentally imported from
that first `.bazelprojct`).

Logic to go looking for anything with a `.bazelproject` suffix is
reasonable, but it should probably still prefer a file literally called
`.bazelproject` if it exists.