fix: show more precise signature help types
The non-resolved signature in signature help needs to rely on the symbol
information type for producing the right signature (as opposed to just
looking at the call directly in the AST). However, this means that some
care must be taken to show a specific type when the member if being seen
from a type which can help further constrain the signature.

(cherry picked from commit da6196d7c4c00f5985065f389ecb3126258db612)