fix: don't show byname hints for defaulted args
This makes sure we don't display an extra `=> ` for default byname
arguments which the user never wrote. Ideally we would have a better way
to check for default arguments, but it looks like like scalac only
fairly recently added an annotation info to indicate that an apply
argument is defaulted. Checking for an empty position is a fallback that
should work on old versions as well.