feat: read the Mill version from YAML frontmatter of the buildfile
With Mill 0.13/1.0, Mill support YAML frontmatter to configure the preferred Mill version.

```scala
//| mill-version: 1.0.0
```

This pull request add support for reading the Mill version from the YAML frontmatter in the main build script.

This change maintains the following precedence rules:

1. if exists, read `.mill-verison`
2. if exists, read `.config/mill-verison`
3. if exists, read YAML frontmatter of `build.mill`
4. if exists, read YAML frontmatter of `build.mill.scala`
5. if exists, read YAML frontmatter of `build.sc`
6. if exists, parse the `DEFAULT_MILL_VERSION` attribute from a local `mill` script

(Steps 3-5 are newly added in this PR.)

Additional, the `-jvm` suffix is stripped from the version, since it is used to denote, that Mill should not attempt to use a native launcher.

The automated unit tests checking proper version parsing was extended to cover the YAML parsing accordingly.