Move background compile on the background thread
We need to use a single thread when calling into the compiler, and the rest of Metals is using a custom CompilerJobQueue. While this is not how the PC should be used, in the short-term it is easier to move compilation there than to make all the calls in the codebase use correct ask* calls.

This requires some duplication of compiler code because many methods are private, including an ugly reflective call. We should eventually move to the supported API.