Prompt for imports (#13)
This PR does two things:

* On `didFocus` events coming from the extension, we check if the file
is part of a target we know about. If not, we send a message to user
asking if we should add it to our targets. We do this simply by calling
`expand(path)`, which sends a request to the BSP for the targets. The
expectation is that the BSP will eventually send a notification
`didChange` notification for us with the new target.
* When we connect to the BSP, metals currently cascades compiles (i.e.
compiles the target and all its reverse dependencies) all open files.
This can be very expensive if you have a file like //common open. So
instead we just compile the target.