[PLAT-134366] Force lazy completion stream to avoid race conditions (#14)
CompletionProvider returns a `Seq`, which under the hood was implemented
as a Stream, which is a lazy data structure. As completions were
returned back to VSCode, the Stream was being forced by the LSP thread,
and each completion element was discovered and built "on the fly",
invoking compiler code on a random thread. This in turn caused race
conditions in the type checker, manifesting as no completions and
sometimes no type errors and plenty of stack traces in the logs. By
forcing the stream to a List we evaluate all elements eagerly before
returning from this method, which makes sure all compiler invocations
happen on the compiler thread.

This is not the only race condition in the completion code path, but it
gets us a bit farther.

Co-authored-by: Iulian Dragos <jaguarul@gmail.com>