Actually respect buildOnFocus (#15)
Turns out `maybeCompileOnDidFocus` is called on two places, on both
`didOpen` and `didFocus`. We only patched `didOpen` to respect
`buildOnFocus`, and we missed the case where `didFocus` would call it.
We now stop chasing callers and just respect the setting on the method
itself.

In case we don't want to build on focus, we either return
AlreadyCompiled or NoBuildTarget in case we don't know about it.