Disable building all open files automatically (#17)
In order to provide a more obvious experience, let's not build things on
the background depending on the file you have open when you reload the
window. Having to click import is not great, but at least has very
obvious user intent.

Right now if you reload the window with a file you don't wish to import,
you are basically forced to import it.