Do not show import notifications if an import is in progress (#18)
This PR is simple: if an import build is happening, we avoid showing
imports.
This is mostly to avoid bombarding users with many potential no-op popus
(likely the first import will be enough for many files you work with).

Unfortunately I couldn't figure out how to dismiss notifications at all.

Also there is a gap between "Discovering targets" and "Import build" --
if the lock gets acquired in between those two, we send notifications
(and it seems to happen in practice).

I want to increment this into a point where we could actually give user
clear progress that something is happening from the moment they click
until the last command executes and control the notifications
accordingly.

---------

Co-authored-by: Iulian Dragos <jaguarul@gmail.com>