improvement: Return a better LLM compatible result from find-dep (#7518)
Previously, we would just return strings and even if a potentially full name of library was set we wouldn't return the version. This confused agents and they though the version is part of the name that got returned.

Now:
- we return explicit result which says what was completed
- try to first complete the next step of the dependency string and only then fallback

The latter means that if for example we complete cats-core name, we will first try `org.typelvel:cats-core:` and if nothing is found complete `org.typevelel:cats-core`.