fix: signature help inside annotation trees (#7579)
Signature help was never triggering anywhere in annotations (top-level
as well as in nested apply calls).

It turns out that the reason was just that the `tree: Tree` position
does not extend to also include the position of the annotations on it.

(cherry picked from commit e727664b144edf8ed9db8270287cb878f9d93e7f)