bugfix: Don't try to find definition for invalid symbols (#7585)
Previously, a lot of reports might show up for users and this was most likely caused by the heuristic not managing to find a proper symbol. Now we validate if what we are looking for is actually a symbol.

I also noticed that we sometime might produce an empty symbolPrefix and try to attach something to it, which was the cause of some of the false positive reports.