refactor: add support for claude
This PR also sort of switches the wording around from being editor
focused to client focused allowing for other future configurations to be
there for other types of clients, like CLIs. This also introduces a new
user configuration option `mcpClient` that a user can manually pass in a
client irregardless of what editor they are working in. This is the
workflow that I use locally setting this to `claude` and then getting my
config (project scoped) automatically generated so that when I start
`claude` everything just works.