feat: discover main classes in dependencies (#7590)
feat: discovery main classes in dependencies

This updates `discovery-jvm-run-command` to match the existing debug
discovery logic fallback of looking in dependencies when a main class
was specified but not found in the build target itself.

One net outcome: making a "run" launch config in VS code will now
uniformly launch the class _not_ in the debug console, regardless of
whether the main class is in a dependency or in the target. (Relatedly,
[this change][0] to `metals-vscode` is probably no longer as important)

[0]: https://github.com/scalameta/metals-vscode/pull/1597