bugfix: Recreate mcp json configuration after restart (#7595)
* bugfix: Recreate mcp json configuration after restart

Previously, the editor specific entry would stay after the mcp server got shutdown and the next time it would not get detected properly as the server would start after the editor would read the entry.

Now, we delete the editor specific file and save the port in the default location to read from later. This seems to work much better without the need to refresh the server.

* improvement: Don't delete the full json mcp configuration

* Only check if config is empty.

Co-authored-by: Katarzyna Marek <kasia@marek.net>

---------

Co-authored-by: Katarzyna Marek <kasia@marek.net>