bugfix: Remove stale top levele symbols from current run
Normally, symbols are invalidated between runs, but we seem to be always at runId = 1, which means no symbols are invalidated.

To fix that, when updating compilation unit I go over old top level symbols and invalidate if they are stale. I also had to fix the isStale check, which would accept stale Symbols such as prefixes of the current symbol.