bugfix: Use the current version when publishing docs (#7633)
Quoting sonatype:

```
Due to a bug with the underlying service we rely on to provide SNAPSHOT hosting, we've had to temporarily remove browse access for SNAPSHOT releases. This is related to an outage we've experienced in early June 2025. You should still be able to publish and consume SNAPSHOT releases as usual, but you cannot browse them via the UI. Please let us know if you encounter any issues.
```