[PLAT-139599] Allow bsp connections with no build targets (#41)
Build servers ad stored in a map <build target> -> <server>. This means that when a build server has not targets, it won't be stored at all, and any subsequent updates that create new target won't be correctly stored. Update the code to keep the list of build server separate from the list of targets.