feat: introduce a format-file MCP tool (#7659)
The specific use-case I have for this is that when using claude code
you're allowed to define tools as hooks. So when you make edits to a
file you can run a hook after it. Often this is something like
formatting. While we can call scalafmt directly I thought it'd be nice
to just add this as another tool as everything is mostly there already.