feat: add in `testName` to the test tool (#7665)
I've seen claude do this a few times with sbt where it uses --z to
target a specific test that it knows it testing what i needs, not a
whole suite. I thought it'd be nice for the test tool to also be used
for this. So not not only can you pass in the testFile, but also the
testName in that file.

(cherry picked from commit ec4d7afca08ea19847fd5296f23c35620e7001bd)