Fix missing diagnostics.  (#46)
The diagnostics object is using a default presentation compiler setting
that does not reflect user settings. After fixes in #25 diagnostics were
no longer displayed.

Changed the default value for PC diagnostics to false in tests, as this
interfered with some test that depend on build errors (certain classes
of errors are not reported by the presentation compiler)

---------

Co-authored-by: Iulian Dragos <jaguarul@gmail.com>