[PLAT-140771] Pass source path to the presentation compiler (#49)
The sourcepath allows the compiler to find types that are not compiled
to class files, but for which we have sources. It assumes the same
package/directory mapping as for class files, so it's not bullet proof
-- but it is the first step towards working on codebases that have not
been built, or new types have been added since the last successful
build.

The main drawback right now is that the source path is based on BSP
source items, which generally are directories, but may not be the case
in Bazel. That work is tracked in
https://databricks.atlassian.net/browse/PLAT-140772

---------

Co-authored-by: Iulian Dragos <jaguarul@gmail.com>