[PLAT-142448] Terminate sync when there are no changes (#55)
When we receive a sync request, we send a inverseSource request to the bsp and set the status as syncing. We then wait for a buildTargetDidChange notification to start importing the project again and finally set the status to synced. However, if the bsp sends no changes, the status remains stuck at "Syncing". Here we change that to reset the status based on whether a target has been discovered or not.