Publish via GH Actions (#50)
This PR sets up a workflow for publishing to GH Package registry, a
simple maven repository that's private to our repo. Most of the work can
be reused when publishing somewhere else:

- set up sbt-dynver for dynamic version numbers based on git tag, sha,
etc. If there is a tag, the version patch is going to be exactly the
tag, otherwise it computes it based on the latest git sha
- setup publishing via github package registry for the same version
numbers as when publishing locally
- update the GH Actions workflow to run our command instead of the real
Metals ci publish step

---------

Co-authored-by: Iulian Dragos <jaguarul@gmail.com>