Add sbt-bloop as a build dependency to simplify using Bloop as the build server (#62)
I've had some issues connecting to the sbt build server and I've found
that Bloop just works better. I think some of the issues are related to
the fact that I use local builds of Metals from this fork where I
haven't published sbt-metals.

This PR adds the sbt-bloop plugin as a direct dependency to the build so
that everything should work correctly as long as you pick Bloop as your
build server and manually run `bloopInstall` from the sbt shell when you
want to update the build structure. I picked "Don't show again" when
Metals asks to import the build.

Co-authored-by: Olafur Geirsson <olafurpg@gmail.com>