Databricks-only: Replace Metals own file watcher with LSP file watching (#59)
Related https://databricks.atlassian.net/browse/PLAT-142718

The goal of this commit is to speed up the indexing step when Metals
re-connects with the build server by eliminating file watching
altogether. In some cases, registering file watchers is taking up to 30
seconds, and it's on the blocking path for making Metals "ready".

Instead of using our own file watcher that listens to precise BSP
directories, we rely on the file watcher in VS Code instead and register
for the `**.{scala,java}` pattern. This more or less takes us back to
the original behavior of Metals, see

https://github.com/scalameta/metals/commit/17b47eebba12f682fa081d3010809a794f440fb9
Simplified, the main reason we stopped using LSP file watching was
because it was unreliable across different editors (VS Code, Neovim,
...). However, for this Metals fork, we only support Cursor (aka. VS
Code) so the LSP file watcher should be reliable enough.

I manually validated that we receive correct `*.{scala,java}` file
watching notifications from Cursor.

NOTE:
Before merging, I want to manually test that things more or less still
work as expected in the monorepo.

---------

Co-authored-by: Olafur Geirsson <olafurpg@gmail.com>