[PLAT-142718] Do not follow symlinks by default. (#61)
Verified with a profiler that almost 2 min of "indexing time" was
actually spent in following symlinks (OS calls such as realpath and
exists). Paths coming from bazelbsp are already normalized and absolute,
so we don't need to repeat this operation. I didn't notice any adverse
effects.

This change shaves off about 80 seconds in the repro linked to the
ticket.