[PLAT-142718] Add a fast path to expireSymbolDefinition (#68)
When indexing workspace sources, one step is to remove cache entries for
scaladoc comments. Initially, the cache is empty and this operation is a
no-op, but the implementation needs to parse sources in order to find
out what top-level definitions should be removed. This commit
short-circuits the parsing step for empty cache, halving the time to
index sources (8s -> 3.5s)