Reimplement repo-wide workspace/symbol based on git instead of BSP (#70)
Previously, workspace/symbol only searched for symbols of synced targets
via BSP. This meant that the feature didn't work at all until you
established a BSP connection, and until Metals had finished indexing the
synced targets.

Now, Metals implements workspace/symbols with a new index that is built
directly from reading Scala/Java files that are checked into Git. It
uses a new LMDB persistent key-value store to make this indexing step
incremental so that it's reasonably fast even in a large monorepo.

This change makes Metals workspace/symbol behave more similarly to other
LSP servers
like tsserver that only search for symbols that defined in the
*workspace* (not from external deps like node_modules/). Users who want
the old behavior can still access it through the custom "Metals: Search
symbol in workspace" command, which we should probably rename to
"Metals: Search Symbol From BSP" or something like that.

## Test Plan
There are e2e integration tests that run the indexer against an actual
git repo to assert the indexing is truly incremental, it invalidates
cache entries from deleted files, etc.

The manual testing plan for me is to add the following
`.vscode/settings.json`
```jsonc
  "metals.serverVersion": "1.5.1-SNAPSHOT",
  "metals.serverProperties": [
    // The final index is "only" ~300mb in memory, but the indexing step can use up more memory temporarily.
    // Metals must have more than 2gb ram even if only temporarily, we're dealing with a large build.
    "-Xmx6G",
    "-XX:+UseZGC",
    "-XX:ZUncommitDelay=30",
    "-XX:ZCollectionInterval=5",
    "-XX:+IgnoreUnrecognizedVMOptions",
    "-Dmetals.statistics=workspace-symbol",
    "-Djol.magicFieldOffset=true",
    "-Djol.tryWithSudo=true",
    "-Djdk.attach.allowAttachSelf",
    "-Dmetals.workspace-symbol-provider=mbt",
    "--add-opens=java.base/java.nio=ALL-UNNAMED",
    "--add-opens=java.base/sun.nio.ch=ALL-UNNAMED",
  ],
```

The `--add-opens` settings are necessary for LMDB to work, and we have
to update metals-vscode to always include them. The jol settings are
needed if you enable `-Dmetals.statistics=memory`, which I used to
confirm that the index consumes ~270mb of memory to hold the git OIDs
and the bloom filters.

After publishing locally, open Metals in the universe repo in Cursor,
open the integrated terminal and run
```
$ tail -f .metals/metals.log | grep workspace/symbol
```
The command above shows logs related to indexing
```
$ tail -f .metals/metals.log | grep workspace/symbol
2025.09.11 13:27:06 INFO  workspace/symbol indexing started
2025.09.11 13:27:08 INFO  time: workspace/symbol discovered 85,391 files in 2.08s
2025.09.11 13:27:10 INFO  time: workspace/symbol 0 cold files, 85391 hot files, discovered in 1.17s
```
Use "Go to Symbol in Workspace" after it has completed indexing and
confirm that it searches the full repo, Java (classes) and Scala
(classes, vals, defs, vars).

## Release Plan

This feature is disabled by default. To start with, users have to add
the server property `-Dmetals.workspace-symbol-provider=mbt` along with
the `--add-opens` and `-Xmx` settings above, and reload Cursor. This
allows us to slowly roll out the functionality, first test it with the
team before enabling by default for the entire organization.

We need to update metals-vscode to include the necessary VM options by
default before we do a bigger rollout.

## Known Limitations - Followups

- We only index the workspace once when the server starts. We should add
a custom `metals/git/didChangeCurrentCommit` notification that triggers
a re-index when the user commits or rebases, etc.
- Incorporate dirty files into the index by listening to
`textDocument/didChange`
- Improve ranking. For example, we could use the imports in the current
file to boost symbols with matching package names.

---------

Co-authored-by: Olafur Geirsson <olafurpg@gmail.com>