Correctly resolve mtags with COURSIER_REPOSITORIES and credentials (#85)
Previously, when the Metals was started with the environment variable
`COURSIER_REPOSITORIES`, Metals would correctly parse the repos and
associate credentials from `~/.config/coursier/credentials.properties`
*but* resolution still failed because we used `.addRepositories()`
instead of `.withRepositories()`. The root problem is that Coursier adds
the same repos but with no credentials and they appeared at the front of
the list. Now, we flip the order so that our custom repos with valid
credentials appear at the front of the list allowing the resolution to
work as expected.

Here are the steps I used to reproduce the issue,
```diff
--- a/build.sbt
+++ b/build.sbt
@@ -538,7 +538,7 @@ lazy val metals = project
     buildInfoKeys := Seq[BuildInfoKey](
       "localSnapshotVersion" -> localSnapshotVersion,
-      "metalsVersion" -> version.value,
+      "metalsVersion" -> "1.5.1-DATABRICKS-8-8-d26262d7", //  version.value,
       "mdocVersion" -> V.mdoc,
       "bspVersion" -> V.bsp,
```

This diff allows you to run a local build of Metals but it will try to
resolve mtags from the remote. I had these settings.json
```jsonc
  "metals.serverVersion": "1.5.1-SNAPSHOT",
  "metals.customRepositories": [
    "central",
    "ivy2local",
    "https://maven.pkg.github.com/REDACTED_ORG/metals"
  ],
```

Co-authored-by: Olafur Geirsson <olafurpg@gmail.com>