fix: go to definition inside of SAM lambda (#7819)
When there's a SAM lambda, semantic DB adds a symbol occurrence for the
extra anonymous class. That occurrence has range corresponding to the
entire lambda. The problem is that when trying to go-to-definition for
something inside of the SAM, `positionOccurrence` might incorrectly
select the anonymous class occurrence as the one whose definition is
being sought...

The tactical fix here is to continue scanning for all occurrences
enclosing the cursor position (instead of just the first) and when an
occurrence is enclosed by another, we prefer it.

The proper fix is probably for the anonymous class occurrence to be a
semantic DB synthetic instead. That symbol never show up in the source
code, so it feels wrong for it to be an occurrence with a range.

See the regression test added for exact case to trigger this.

Fixes #7797