[PLAT-142718] A couple of minor optimizations (#83)
These together bring down the indexing/importing time close to 15s for a
large project. Further optimization is possible but we reached the point
of diminishing returns. Each of the items below amount to a couple of
seconds for the repro in the ticket, but I don't have individual
benchmarks.

- avoid expensive URI parsing for file URIs
- avoid expensive Json parsing by caching type adapters for maven
dependency info
- avoid expensive TrieMap updates when the value didn't change while
building inverted indexes for source jars

Example log from importing the repro project:
```
2025.09.19 11:02:12 INFO  time: indexed library sources for main in 7.52s
..
2025.09.19 11:02:13 INFO  time: indexed workspace in 15s
```

---------

Co-authored-by: Olaf Geirsson <olaf.geirsson@databricks.com>