Fix resolution of local builds (#88)
Previously, local 1.5.1-SNAPSHOT builds failed on the latest canary
image in arca because we set the repo `ivy2Local` instead of `ivy2local`
(all lowercase). I'm not sure where exactly Coursier compares it in a
case-insensitive way because I copied the original code from how it
parses repos, but I suspect it's somewhere in the bootstrap launcher.
Either way, we now handle this special case.

Co-authored-by: Olafur Geirsson <olafurpg@gmail.com>