Move mbt behind a feature flag (#90)
Previously, the mbt-based workspace symbol solution was gated behind a
system property `-Dmetals.workspace-symbol-provider=mbt`. Now, it's a
user-setting that can be edited in `settings.json`, which Metals
dynamically reacts to.
```jsonc
  "metals.workspaceSymbolProvider": "mbt", // or "bsp"
```
Also, if the config is not set, then Metals now defaults to mbt if the
`MBT_WORKSPACE_SYMBOL_PROVIDER` feature flag is enabled.

Also:
- If the process does not have the necessary VM options to use LMDB,
Metals now logs a helpful warning message with instructions on how to
fix the problem, and it defaults to bsp.
- The old system property still works thanks to how Metals parses
config.

## Testing plan

- With no custom settings or VM options, it defaults to bsp and logs a
warning
- With correct VM options and you're in the feature flag, it defaults to
mbt
- You can always explicitly override the setting
- Metals dynamically reacts to updates to the setting (it will even
re-index with mbt when it changes)

---------

Co-authored-by: Olafur Geirsson <olafurpg@gmail.com>