[Sync] Do not compile after syncing (#108)
`compileFile` is guarded by `userConfig.buildOnFocus` everywhere but here. This causes the compilation to start again after the indexing has finished. In most cases this is fast because the focussed file is the same as the synced one or part of the same target. But otherwise this will trigger a compilation that we don't really need.