Fix error when using older Scala versions (#113)
The PR https://github.com/REDACTED_ORG/metals/pull/105 had a
regression where Metals was still trying to resolve an old version of
mtags when the user had an old version of Scala. I had tested it
manually before sending that PR and it seemed to work fine, but it was
failing for me now when I did one more test after we merged the PR.

Now, we forcefully bump up the Scala version for the mtags classpath to
match the Scala version that mtags was built against. I manually tested
this in an sbt build with the config
```scala
inThisBuild(List(
  scalaVersion := "2.13.10",
  javaHome := Some(file(sys.env("JAVA_HOME"))),
))

lazy val lib = project

lazy val app = project
.dependsOn(lib)
.settings(libraryDependencies += "junit" % "junit" % "4.13.2")
```

Previously, it would not provide any completions, it errored with
"Unsupported Scala VERSION". Now, everything works as expected.

Co-authored-by: Olafur Geirsson <olafurpg@gmail.com>