bugfix: Readd build tool when missing
It seems if anyone removed their .metals directory, but left everything else, we would never set the used build tools again.
This meant that when we check for sbt, we wouldn't get any result and we wouldn't support build.sbt file.

Now, if build tool is missing we requery it.