Power WorkspaceSymbolProvider with mbt symbol search (#118)
Previously, if you enabled the `workspaceSymbolProvider: "mbt"` setting,
then it would only use mbt to power `workspace/symbol` requests while
the old non-mbt implementation still powered other non-workspace/symbol
functionality via WorkspaceSymbolProvider. For example, auto-import
completions and "Add import" code actions rely on the old
`WorkspaceSymbolProvider`.

Now, when you enable `workspaceSymbolProvider: "mbt"`, you also enable
it for all the other functionality. This is a blocking step for the
upcoming Java support to import any symbol in the repo via -sourcepath.
Without this PR, you could still import any symbol but you'd have to
type out the import manually. With this PR, you can auto-import any
symbol in the repo.

Important limitation: mbt still doesn't support BSP buildTarget-based
symbol search so we still fallback to the non-mbt index for those
queries. This change makes this PR largely a no-op, it's only preparing
the stage for the Java PR that will hit this API with an empty
buildTarget to power auto-import completions.

## Test Plan

The biggest risk in this PR is introducing a regression for Scala
auto-import completions. I manually confirmed that it's working as
expected (it's BSP buildTarget-based) by syncing two files

- `experimental/iulian.dragos/class-indexer/src/ClassIndexer.scala`:
where I can auto-import symbols `scala.reflect.internal.util.NoPosition`
because the target depends on scala-compiler, but I can't import
`com.fasterxml.jackson.databind.JsonNode` because it's not on the
classpath.
- `example/Example.scala`: opposite
of the symbols above

Another user-facing behavioral change is that "Go to symbol in
workspace" now falls back to the classpath when you have enabled mbt and
there's no match in the workspace. Before, it only searched through
workspace symbols. For example, test it with the query
`BasicFileAttributes`, which has no matches in the repo but returns the
JDK symbol with this name (after this PR, no results with latest stable
if you enabled mbt)

---------

Co-authored-by: Olafur Geirsson <olafurpg@gmail.com>