Use BatchedFunction to debounce diagnostics (#114)
Previously, we used a custom class to debounce diagnostics. Now, we
achieve the same effects using the `BatchedFunction` helper class that
we are using elsewhere in the codebase. The debouncing logic is a bit
different, we always flush every 500ms even if you are still typing.
Before, we flushed only once you stopped typing.

Also, by moving this new solution, I've updated the
`didChange`/`didOpen`/`didFocus` handler to return futures that await on
the presentation compiler diagnostics. This means that we no longer have
to add Thread.sleep in tests when asserting that the PC has published or
cleared diagnostics.

## New Behavior
Diagnostics are flushed every 500ms.


https://github.com/user-attachments/assets/ab088c07-e1d7-49e0-a132-4e960d7f7906



## Old Behavior
Diagnostics would only be flushed after 500ms of inactivity.




https://github.com/user-attachments/assets/db70b870-3537-48bd-84bf-b739f7238136

Co-authored-by: Olafur Geirsson <olafurpg@gmail.com>