Make it easier to debug SemanticDB issues (#125)
> Extracted from the Java PR.

Previously, there was no good way to visualize the SemanticDB data that
the compiler sees. There exists a "Metals: Show SemanticDB Detailed"
command but it uses a cryptic ~10yo format from `metap` that I never
found helpful. Now, I've replaced this format with a much clearer output
that find much simpler to view. This is the same format we have been
using to develop SCIP indexers at Sourcegraph for the past five years,
and it's the format we have used to test the semanticdb-javac compiler
plugin.

## Examples

See test cases for concrete examples. 

**Before**
```
package a

object O/*a.O.*/ {
  val a/*a.O.a.*/ = 123
  val b/*a.O.b.*/ = a/*a.O.a.*/ +/*scala.Int#`+`(+4).*/ 1
}
```

**After**
```
   package a
//         ^ definition a/

   object O {
//        ^ definition a/O.
     val a = 123
//       ^ definition a/O.a.
     val b = a + 1
//       ^ definition a/O.b.
//           ^ reference a/O.a.
//             ^ reference scala/Int#`+`(+4).
```

The problem with the old format is that lines end up having a lot of
identifier and everything shifts horizontally to include the comments.
The old format also had the problem that it make syntax highlighting
look very weird while the new format makes it really easy to see the
original code with syntax highlighting enabled. The new format also
doesn't have strange issues when the symbols themselves contain syntax
like `*/` to close an inline comment.

## Risk

This is a low-risk PR because it only impacts debugging command "Show
SemanticDB Detailed"

Co-authored-by: Olafur Geirsson <olafurpg@gmail.com>