Make internal fork work with fresh sbt builds (#121)
Previously, if you opened our internal Metals fork in a random sbt
project then it will fail because it tries to add an sbt-metals
dependency to `project/plugins.sbt` with the matching version name as
Metals itself and we don't publish sbt-metals. Publishing a fork of
sbt-metals isn't going to work with our custom publish infrastructure
for a variety of reasons.

- sbt needs our custom resolver to pick it the artifact from
- our publish job assumes publishLocal publishes normal maven
  coordinates while sbt plugins use a custom ivy layout

Instead, of getting that working, it's simpler to just generate
`project/metals.sbt` files that resolve the latest matching `sbt-metals`
plugin, which is the version "1.5.1".

Co-authored-by: Olafur Geirsson <olafurpg@gmail.com>