Enable telemetry by default when Metals is instrumented (#127)
Previously, you had to complete two steps to enable telemetry:

- Instrument the Metals classpath with SPI (Service
  Provider Interface) implementations for `FeatureFlagProvider` and
  `MonitoringClient`.
- Add the system property `-Dmetals.telemetry=enabled`.

This was problematic because it was easy to fail the second step if a
user manually customized the `metals.serverProperties` setting. We rely
on telemetry for feature flags so these users would not get mbt-based
symbol search even if the feature flag is enabled for their username.

Now, you only need to complete the first step to enable telemetry by
default. Users can still disable telemetry with
`-Dmetals.telemetry=disabled`. In OSS Metals, this should still be a
reasonable default because Metals won't be instrumented with SPIs out of
the box.

Co-authored-by: Olafur Geirsson <olafurpg@gmail.com>