Improve error handling in Doctor (#129)
Previously, Metals ran the Doctor during the indexing step, and it
didn't handle errors from the Doctor. We had a bug report from a user
where the JAVA_HOME environment variable pointed to a non-existing
directory so it errored and broke indexing. Errors in the doctor should
not be fatal so this PR adds a few try/catch clauses to not let failures
in the Doctor propagate into other systems.

Co-authored-by: Olafur Geirsson <olafurpg@gmail.com>