Inline semanticdb-javac compiler plugin from sourcegraph/scip-java repo (#128)
We will heavily rely on this plugin for the upcoming Java support and we
need to be able to make core changes to expose more APIs and fix errors
that happen more frequently in broken code, which is comming in the IDE.

The upstream repo is probably not going to get maintained much anyways
so I think it makes sense that the Scalameta organization takes over
long-term maintenance of it. We already publish
`org.scalameta:semanticdb-scalac` so it makes sense that the same org
publishes `org.scalameta:semanticdb-javac`.


## Risk

None. This is a no-op PR besides the renaming of the package for the
Java SemanticDB bindings.
The semanticdb-javac plugin isn't used yet, it's just copy-pasted code.

Co-authored-by: Olafur Geirsson <olafurpg@gmail.com>